Version 2.1.0
The virtual pet game now works using an embedded database and no longer replies on a text file to store player's data. Instead, data will be read and written to an embedded database. The only exception is whens storing the math questions, the math questions is read in from a file and written to the database.